move EthFlow to cow-react
