+++ src/cow-react/modules/ethFlow/containers/EthFlowModal/utils/getDerivedEthFlowState.ts
