+++ src/cow-react/modules/ethFlow/pure/EthFlowModalContent/configs.ts
